requirements.txt
pillow
matplotlib
pyyaml
scipy
scikit-image
joblib
yaml
imageio